Split out an "Error" module.
